+++ src/main/java/doctor_m/network/ConfigOpenPacket.java
@@ -1,0 +1,7 @@
+package doctor_m.network;
+
+import net.minecraft.util.Identifier;
+
+public class ConfigOpenPacket {
+    public static final Identifier ID = new Identifier("doctor_m", "open_config");
+}